Fix revision types in new replication stuff.

+++ arangod/RocksDBEngine/RocksDBTransactionCollection.cpp
@@ -260,13 +260,13 @@ void RocksDBTransactionCollection::commitCounts(TRI_voc_tid_t trxId, uint64_t co
 
 void RocksDBTransactionCollection::trackInsert(TRI_voc_rid_t rid) {
   if (_collection->syncByRevision()) {
-    _trackedOperations.inserts.emplace_back(static_cast<std::size_t>(rid));
+    _trackedOperations.inserts.emplace_back(rid);
   }
 }
 
 void RocksDBTransactionCollection::trackRemove(TRI_voc_rid_t rid) {
   if (_collection->syncByRevision()) {
-    _trackedOperations.removals.emplace_back(static_cast<std::size_t>(rid));
+    _trackedOperations.removals.emplace_back(rid);
   }
 }
